dom/tests/mochitest/orientation/test_bug507902.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=507902
-->
<head>
  <title>Test for watchPosition </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=507902">Mozilla Bug 507902</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

function boom()
{
  window.addEventListener("unload", function(){}, false);
  window.addEventListener("devicemotion", function(){}, false);
  location = "data:text/html,2";

  ok(1, "leak will be at the end of mochitests. so pass.");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

window.addEventListener("load", function() { setTimeout(boom, 0); }, false);

</script>
</pre>
</body>
</html>
